requirements.txt
bcrypt==4.1.1
cffi==1.16.0
cryptography==41.0.7
decorator==5.1.1
Deprecated==1.2.14
Fabric3==1.14.post1
invoke==2.2.0
paramiko==2.12.0
pycparser==2.21
PyNaCl==1.5.0
six==1.16.0
wrapt==1.16.0
